+++ env0/provider.go
@@ -54,9 +54,9 @@ func Provider(version string) plugin.ProviderFunc {
 				"env0_configuration_variable":  dataConfigurationVariable(),
 				"env0_template":                dataTemplate(),
 				"env0_ssh_key":                 dataSshKey(),
-				"env0_aws_cost_credentials":    dataAwsCostCredentials(),
-				"env0_azure_cost_credentials":  dataAzureCostCredentials(),
-				"env0_google_cost_credentials": dataGoogleCostCredentials(),
+				"env0_aws_cost_credentials":    dataCostCredentials(string(client.AwsCostCredentialsType)),
+				"env0_azure_cost_credentials":  dataCostCredentials(string(client.AzureCostCredentialsType)),
+				"env0_google_cost_credentials": dataCostCredentials(string(client.GoogleCostCredentialsType)),
 				"env0_aws_credentials":         dataAwsCredentials(),
 				"env0_gcp_credentials":         dataGcpCredentials(),
 				"env0_azure_credentials":       dataAzureCredentials(),
